requirements.txt
streamlit>=1.32.0
pytesseract>=0.3.10
Pillow>=10.0.0
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
pandas>=2.0.0
opencv-python-headless>=4.9.0
nltk>=3.8.0
PyMuPDF>=1.24.0
plotly>=5.20.0
numpy>=1.26.0
Flask>=3.0.0
